+++ src/Geopilot.Frontend/src/components/visualizations/map/mapVisualization.tsx
@@ -25,9 +25,9 @@ import WMTS, { optionsFromCapabilities } from "ol/source/WMTS";
 import { Circle, Fill, Stroke, Style } from "ol/style";
 import View from "ol/View";
 import proj4 from "proj4";
-import { MapLayer, MapVisualizationConfig } from "../../../../api/apiInterfaces";
-import { IconButton } from "../../../../components/buttons";
-import { stopStepSwipePropagation } from "../../../../hooks/useStepSwipe";
+import { MapLayer, MapVisualizationConfig } from "../../../api/apiInterfaces";
+import { IconButton } from "../../../components/buttons";
+import { stopStepSwipePropagation } from "../../../hooks/useStepSwipe";
 import { LayerSwitcher, LayerSwitcherProperties } from "./layerSwitcher";
 import { MapVisualizationContext } from "./mapVisualizationContext";
 import "ol/ol.css";
