+++ src/Geopilot.Frontend/src/components/visualizations/map/mapVisualizationContext.tsx
