+++ src/Geopilot.Frontend/src/components/visualizations/map/mapVisualizationProvider.tsx
