+++ src/Geopilot.Frontend/src/components/visualizations/visualizationRegistry.tsx
